fix: VehiclesRentalsPricesAllEntity PriceRent type

+++ ArkanisOverlay/Data/Entities/UEX/VehiclesRentalsPricesAllEntity.cs
@@ -6,7 +6,7 @@ public class VehiclesRentalsPricesAllEntity : BaseEntity
 {
     [JsonPropertyName("id_vehicle")] public int IdVehicle { get; set; }
     [JsonPropertyName("id_terminal")] public int IdTerminal { get; set; }
-    [JsonPropertyName("price_rent")] public float PriceRent { get; set; } // last
+    [JsonPropertyName("price_rent")] public decimal PriceRent { get; set; } // last
     [JsonPropertyName("vehicle_name")] public string VehicleName { get; set; }
     [JsonPropertyName("terminal_name")] public string TerminalName { get; set; }
 }
